+++ src/com/pinktwins/elephant/CustomEditor.java
@@ -69,6 +69,7 @@ import com.pinktwins.elephant.data.Note;
 import com.pinktwins.elephant.eventbus.StyleCommandEvent;
 import com.pinktwins.elephant.eventbus.UndoRedoStateUpdateRequest;
 import com.pinktwins.elephant.ui.AutoIndentAction;
+import com.pinktwins.elephant.ui.HomeAction;
 import com.pinktwins.elephant.ui.ShiftTabAction;
 import com.pinktwins.elephant.ui.TabAction;
 import com.pinktwins.elephant.util.CustomMouseListener;
@@ -594,6 +595,8 @@ public class CustomEditor extends RoundPanel {
 		// enable Tab and Shift-Tab behavior for bullet lists
 		note.registerKeyboardAction(new TabAction(), KeyStroke.getKeyStroke(KeyEvent.VK_TAB, 0), JComponent.WHEN_FOCUSED);
 		note.registerKeyboardAction(new ShiftTabAction(), KeyStroke.getKeyStroke(KeyEvent.VK_TAB, java.awt.event.InputEvent.SHIFT_DOWN_MASK), JComponent.WHEN_FOCUSED);
+
+		note.registerKeyboardAction(new HomeAction(), KeyStroke.getKeyStroke(KeyEvent.VK_HOME, 0), JComponent.WHEN_FOCUSED);
 
 		maybeImporting = false;